fix: extract header into new comp
